- BIG CHANGE! - generate Model_init.xml instead of Model_init.txt - clean objects in Model.makefile before compiling (Make doesn't seem to get that   Model.c is newer than Model.o if we issue two simulate commands for the same model,   even if the model changed via loadFile.) - check the result of fflush in systemimpl.c and printimpl.c - fflush in c_runtime/linearize.cpp - added c_runtime/simulation_input_xml.h and cpp to read the Model_init.xml file. - link the generated code with the XML parser library -lexpat as libc_runtime needs it now - Model.c and Model_init.xml will now contain a GUID which is used in c_runtime/simulation_input_xml.cpp   to check if the Model_init.xml is matched with Model.c (Model.exe) - The Model_init.xml is based on FMI but with much more information that we need for OMC. - The Model_init.xml file it has no schema yet as it may change in the near future to include   more things that are now static in the generated code (Model.c). This will make compilation   of generated code faster. - disabled testsuite/interactive-simulation/runInteractiveSimulation.mos and added only a   compilation test (more work is needed to make the run test more robust). - updated tests and compiler files with the name change from _init.txt to _init.xml.
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@9222 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ c_runtime/simulation_result_plt.cpp
@@ -101,7 +101,7 @@ void simulation_result_plt::emit()
   if (actualPoints < maxPoints) {
     if(!isInteractiveSimulation())add_result(simulationResultData,&actualPoints); //used for non-interactive simulation
   } else {
-    maxPoints = 1.4*maxPoints + (maxPoints-actualPoints) + 2000;
+    maxPoints = (long)(1.4*maxPoints + (maxPoints-actualPoints) + 2000);
     // cerr << "realloc simulationResultData to a size of " << maxPoints * dataSize * sizeof(double) << endl;
     simulationResultData = (double*)realloc(simulationResultData, maxPoints * dataSize * sizeof(double));
     if (!simulationResultData) {
